chore: setting things up for new home page
